Interpolation1D

ClimaInterpolations.Interpolation1D.interpolate1d! — Function

interpolate1d!(
    ftarget::AbstractArray{FT, N},
    xsource::AbstractArray{FT, NSG},
    xtarget::AbstractArray{FT, NTG},
    fsource::AbstractArray{FT, N},
    order,
    extrapolate = Flat(),
    reverse = false,
) where {FT, N, NSG, NTG}

Interpolate fsource, defined on grid xsource, onto the xtarget grid. Here the source grid xsource is an N-dimensional array of columns. The first dimension is assumed to be the column dimension. Each column can have a different grid. It is assumed that both xsource and xtarget are either mononically increasing (reverse = false) or decreasing (reverse = true).

ClimaInterpolations.Interpolation1D.interpolate_column! — Function

interpolate_column!(
    ftarget,
    xsource,
    xtarget,
    fsource,
    order,
    extrapolate;
    reverse = false,
)

Interpolate fsource, defined on column (1D) grid xsource, onto the xtarget grid. It is assumed that both xsource and xtarget are either mononically increasing (reverse = false) or decreasing (reverse = true). This is a convenience function primarily intended for internal use.

ClimaInterpolations.Interpolation1D.Interpolate1D — Type

Interpolate1D{V, IO, EO}

This struct stores the source grid (xsource), function defined on the source grid (fsource), interpolation order and extrapolation order for 1-dimensional interpolation. This struct is designed to be be used in broadcasting calls for 1-dimensional interpolation.

E.g.:

itp = Interpolate1D( xsource, fsource, interpolationorder = Linear(), extrapolationorder = extrapolation, )

ftarget = itp.(xtarget)

ClimaInterpolations.Interpolation1D.get_stencil — Function

get_stencil(alg::Linear, xsource, xtarget; first = 1, extrapolate = Flat(), reverse = false)

This function returns the starting and ending points, in the source grid xsource, for the stencil needed for linear interpolation. If xtarget is outside the range of xsource, this returns the corresponding stencil needed for extrapolation. The stencil specification is characterized by [st, en] where st is the starting point and en is the ending point of the stencil in the target grid. This function returns the tuple (st, en). Linear and Flat extrapolation schemes are supported at the boundaries. The argument first can be used to speedup the search, by providing a more efficient starting point for the search.

ClimaInterpolations.Interpolation1D.Order1D — Type

Order1D

Abstract type for interpolation schemes.

ClimaInterpolations.Interpolation1D.Extrapolate1D — Type

Extrapolate1D

Abstract type for extrapolation schemes.

ClimaInterpolations.Interpolation1D.Linear — Type

Linear <: Order1D

Use linear interpolation.

ClimaInterpolations.Interpolation1D.Flat — Type

Flat <: Extrapolate1D

Use flat extrapolation.

ClimaInterpolations.Interpolation1D.LinearExtrapolation — Type

LinearExtrapolation <: Extrapolate1D

Use linear extrapolation.